requirements.txt
# Core
python-dotenv>=1.0.0
httpx>=0.27.0
asyncpg>=0.29.0
pydantic>=2.0.0
rich>=13.0.0
openai>=1.0.0

# Auth0 (Secure Agent Identity)
auth0-ai

# Airbyte Agent Connectors (Multi-Source Data)
airbyte-agent-github>=0.18.0
airbyte-agent-slack>=0.1.0

# Overmind (Prompt Optimization)
overmind-sdk>=0.1.0

# Aerospike (Real-time Cache)
aerospike>=19.0.0
